+++ lib/src/_columns.dart
@@ -1,4 +1,5 @@
 import 'package:domino/domino.dart';
+import 'package:domino/helpers.dart';
 
 import '_modifiers.dart';
 
@@ -6,23 +7,21 @@ import '_modifiers.dart';
 
 // TODO: add variable gap (is-1 .. is-9)
 
-Element columns({
-  Mods mods,
-  List content,
+Element columns(
+  content, {
   bool noGap,
   bool multiLine,
   bool centered,
 }) =>
-    apply(
-        new Element('div', content: content)
-          ..addClass('columns')
-          ..addClass(noGapClass(noGap))
-          ..addClass(multiLineClass(multiLine))
-          ..addClass(centeredClass(centered)),
-        mods);
+    div([
+      clazz('columns'),
+      noGap == true ? Modifier.noGap : null,
+      multiLine == true ? Modifier.multiLine : null,
+      centered == true ? Modifier.centered : null,
+      content,
+    ]);
 
-Element column({Mods mods, content}) =>
-    apply(new Element('div', classes: ['column'], content: content), mods);
+Element column(content) => div([clazz('column'), content]);
 
 // TODO: add Basics
 // TODO: add Sizes