+++ lib/src/_components.dart
@@ -1,65 +1,67 @@
 import 'package:domino/domino.dart';
+import 'package:domino/helpers.dart' show div;
 
 import '_elements.dart';
-import '_modifiers.dart';
 
 // https://bulma.io/documentation/components/breadcrumb/
 
-Element breadcrumb({Mods mods, List items}) {
-  final children =
-      items.map((item) => new Element('li', content: item)).toList();
-  children.last.addClass('is-active');
-  return apply(
-      new Element(
-        'nav',
-        classes: ['breadcrumb'],
-        attrs: {'aria-label': 'breadcrumbs'},
-        content: new Element('ul', content: children),
-      ),
-      mods);
+Element breadcrumb({mods, List items}) {
+  final children = [];
+  for (int i = 0; i < items.length; i++) {
+    var item = items[i];
+    if (i == items.length - 1) {
+      item = [clazz('is-active'), item];
+    }
+    children.add(new Element('li', item));
+  }
+  return new Element(
+    'nav',
+    [
+      clazz('breadcrumb'),
+      attr('aria-label', 'breadcrumbs'),
+      mods,
+      new Element('ul', children),
+    ],
+  );
 }
 
-Element card({Mods mods, String title, content}) {
+Element card({mods, String title, content}) {
   final children = [];
   var headerChildren;
   if (title != null) {
     headerChildren ??= [];
-    headerChildren.add(new Element('p', classes: ['card-header-title']));
+    headerChildren.add(new Element('p', [clazz('card-header-title'), title]));
   }
   // TODO: add header icon
   if (headerChildren != null) {
-    children.add(new Element('header',
-        classes: ['card-header'], content: headerChildren));
+    children.add(new Element('header', [clazz('card-header'), headerChildren]));
   }
 
   // TODO: add image
 
   if (content != null) {
-    children
-        .add(new Element('div', classes: ['card-content'], content: content));
+    children.add(div([clazz('card-content'), content]));
   }
 
   // TODO: add footer
 
-  return apply(new Element('div', classes: ['card'], content: children), mods);
+  return div([clazz('card'), mods, children]);
 }
 
-Element message({Mods mods, String title, EventHandler onDelete, content}) {
+Element message({mods, String title, EventHandler onDelete, content}) {
   final children = [];
   if (title != null || onDelete != null) {
-    children.add(new Element('div', classes: [
-      'message-header'
-    ], content: [
-      new Element('p', content: title),
+    children.add(div([
+      clazz('message-header'),
+      mods,
+      new Element('p', title),
       delete(onClick: onDelete),
     ]));
   }
   if (content != null) {
-    children
-        .add(new Element('div', classes: ['message-body'], content: content));
+    children.add(div([clazz('message-body'), content]));
   }
-  return apply(
-      new Element('article', classes: ['message'], content: children), mods);
+  return new Element('article', [clazz('message'), mods, children]);
 }
 
 // TODO: add Dropdown