+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Removed `BulmaComponent`, clients should use `Mods` instead.
 - Removed most `Component` classes, using simple methods instead in simple cases.
+- Simplified parameters (removed `is` prefix).
 
 New features: